Move getTextForSelection into doc helpers module.
