+++ document/abstract_document.js
@@ -7,6 +7,7 @@ var Selection = require('./selection');
 var PropertySelection = require('./property_selection');
 var ContainerSelection = require('./container_selection');
 var TableSelection = require('./table_selection');
+var docHelpers = require('./helpers');
 
 function AbstractDocument(schema) {
   Substance.EventEmitter.call(this);
@@ -147,31 +148,7 @@ AbstractDocument.Prototype = function() {
   };
 
   this.getTextForSelection = function(sel) {
-    var result = [];
-    var text;
-    if (!sel || sel.isNull()) {
-      return "";
-    } else if (sel.isPropertySelection()) {
-      text = this.get(sel.start.path);
-      result.push(text.substring(sel.start.offset, sel.end.offset));
-    } else if (sel.isContainerSelection()) {
-      var container = this.get(sel.containerId);
-      var components = container.getComponentsForRange(sel.range);
-      for (var i = 0; i < components.length; i++) {
-        var comp = components[i];
-        text = this.get(comp.path);
-        if (components.length === 1) {
-          result.push(text.substring(sel.start.offset, sel.end.offset));
-        } else if (i===0) {
-          result.push(text.substring(sel.start.offset));
-        } else if (i===components.length-1) {
-          result.push(text.substring(0, sel.end.offset));
-        } else {
-          result.push(text);
-        }
-      }
-    }
-    return result.join('');
+    return docHelpers.getTextForSelection(this, sel);
   };
 
   this.toJSON = function() {
